requirements.txt
pygame-ce>=2.5.2,<3.0.0
pygame-gui>=0.6.12,<0.7.0
